+++ CHANGELOG.md
@@ -17,6 +17,8 @@ Templates:
 #### Features
 
 #### Improvements
+- (deps) [#41](https://github.com/bcdevtools/consvp/pull/41) Update deps cvp-streaming-core v1.1.0
+- (codec) [#42](https://github.com/bcdevtools/consvp/pull/42) Add new flag `--codec` to specify codec for streaming
 
 #### Bug Fixes